Reorganize unit tests in liveblocks-zustand
